LayoutTests/platform/chromium-mac/http/tests/xmlhttprequest/XMLHttpRequestException-expected.txt

Tests the properties of the XMLHttpRequestException object.

PASS e.toString() is "Error: NETWORK_ERR: XMLHttpRequest Exception 101"
PASS Object.prototype.toString.call(e) is "[object XMLHttpRequestException]"
FAIL Object.prototype.toString.call(e.__proto__) is not "[object XMLHttpRequestException]". Was "[object Object]"
FAIL e.constructor.toString() is not "[object XMLHttpRequestExceptionConstructor]". Was "function XMLHttpRequestException() { [native code] }"
PASS e.constructor is window.XMLHttpRequestException
PASS e.NETWORK_ERR is e.constructor.NETWORK_ERR
PASS e.NETWORK_ERR is 101
